pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flow"
version = "0.1.0"
description = "Personal AI dev harness — cost-optimized Claude Code orchestrator"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.40.0",
    "duckdb>=1.1.0",
    "fastapi>=0.115.0",
    "langfuse>=3.0.0",
    "litellm>=1.50.0",
    "prompt-toolkit>=3.0.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
flow = "flow.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/flow"]

[tool.hatch.build.targets.editable]
packages = ["src/flow"]
